requirements.txt
httpx==0.27.2
langchain_core==0.3.12
langchain_openai==0.2.3
langtrace_python_sdk==3.2.0
pipreqs==0.5.0
psycopg2==2.9.10
pydantic==2.9.2
python-dotenv==1.0.1
python_Levenshtein==0.26.0
PyYAML==6.0.2
reportlab==4.2.5
selenium==4.25.0
webdriver_manager==4.0.2
